Remove API header dependency from mantid.kernel
