+++ Framework/PythonInterface/mantid/api/src/Exports/Workspace.cpp
@@ -8,8 +8,8 @@
 #include "MantidAPI/WorkspaceHistory.h"
 #include "MantidKernel/WarningSuppressions.h"
 
+#include "MantidPythonInterface/api/RegisterWorkspacePtrToPython.h"
 #include "MantidPythonInterface/core/GetPointer.h"
-#include "MantidPythonInterface/kernel/Registry/RegisterWorkspacePtrToPython.h"
 
 #include <boost/python/class.hpp>
 #include <boost/python/copy_const_reference.hpp>
