+++ Framework/PythonInterface/mantid/kernel/CMakeLists.txt
@@ -82,7 +82,6 @@ set(INC_FILES
     inc/MantidPythonInterface/kernel/Registry/SequenceTypeHandler.h
     inc/MantidPythonInterface/kernel/Registry/TypedPropertyValueHandler.h
     inc/MantidPythonInterface/kernel/Registry/TypeRegistry.h
-    inc/MantidPythonInterface/kernel/Registry/RegisterWorkspacePtrToPython.h
 )
 
 # Generate the packagesetup module for the build & install package
