+++ Framework/PythonInterface/mantid/kernel/inc/MantidPythonInterface/kernel/Registry/TypedPropertyValueHandler.h
@@ -6,11 +6,10 @@
 // SPDX - License - Identifier: GPL - 3.0 +
 #pragma once
 
-#include "MantidPythonInterface/core/ExtractWorkspace.h"
+#include "MantidPythonInterface/core/ExtractSharedPtr.h"
 #include "MantidPythonInterface/core/IsNone.h" // includes object.hpp
 #include "MantidPythonInterface/kernel/Registry/PropertyValueHandler.h"
 
-#include "MantidAPI/Workspace.h"
 #include "MantidKernel/IPropertyManager.h"
 #include "MantidKernel/PropertyWithValue.h"
 
@@ -21,9 +20,7 @@
 
 #include <string>
 
-namespace Mantid {
-namespace PythonInterface {
-namespace Registry {
+namespace Mantid::PythonInterface::Registry {
 /**
  * This class provides a templated class object that is able to take a
  * python object and perform operations with a given C type.
@@ -75,71 +72,4 @@ struct DLLExport TypedPropertyValueHandler : public PropertyValueHandler {
   }
 };
 
-//
-// Specialization for shared_ptr types. They need special handling for
-// workspaces
-//
-template <typename T>
-struct DLLExport TypedPropertyValueHandler<
-    std::shared_ptr<T>,
-    typename std::enable_if<std::is_base_of<API::Workspace, T>::value>::type>
-    : public PropertyValueHandler {
-  /// Type required by TypeRegistry framework
-  using HeldType = std::shared_ptr<T>;
-
-  /// Convenience typedef
-  using PointeeType = T;
-  /// Convenience typedef
-  using PropertyValueType = std::shared_ptr<T>;
-
-  /**
-   * Set function to handle Python -> C++ calls and get the correct type
-   * @param alg :: A pointer to an IPropertyManager
-   * @param name :: The name of the property
-   * @param value :: A boost python object that stores the value
-   */
-  void set(Kernel::IPropertyManager *alg, const std::string &name,
-           const boost::python::object &value) const override {
-    if (value == boost::python::object())
-      alg->setProperty<HeldType>(name, std::shared_ptr<T>(nullptr));
-    else
-      alg->setProperty<HeldType>(
-          name, std::dynamic_pointer_cast<T>(ExtractWorkspace(value)()));
-  }
-
-  /**
-   * Create a PropertyWithValue from the given python object value
-   * @param name :: The name of the property
-   * @param defaultValue :: The defaultValue of the property. The object
-   * attempts to extract
-   * a value of type ValueType from the python object
-   * @param validator :: A python object pointing to a validator instance, which
-   * can be None.
-   * @param direction :: The direction of the property
-   * @returns A pointer to a newly constructed property instance
-   */
-  std::unique_ptr<Kernel::Property>
-  create(const std::string &name, const boost::python::object &defaultValue,
-         const boost::python::object &validator,
-         const unsigned int direction) const override {
-    using boost::python::extract;
-    using Kernel::IValidator;
-    using Kernel::Property;
-    using Kernel::PropertyWithValue;
-    const PropertyValueType valueInC =
-        extract<PropertyValueType>(defaultValue)();
-    std::unique_ptr<Property> valueProp;
-    if (isNone(validator)) {
-      valueProp = std::make_unique<PropertyWithValue<PropertyValueType>>(
-          name, valueInC, direction);
-    } else {
-      const IValidator *propValidator = extract<IValidator *>(validator);
-      valueProp = std::make_unique<PropertyWithValue<PropertyValueType>>(
-          name, valueInC, propValidator->clone(), direction);
-    }
-    return valueProp;
-  }
-};
-} // namespace Registry
-} // namespace PythonInterface
-} // namespace Mantid
+} // namespace Mantid::PythonInterface::Registry
